Restructured the organization of the product files
Network Alchemy original started out as a much smaller vision intended
to solve a very specific problem. The original intention was abandoned,
but in the process I recognized the infrastructure that I had built
along the way could be valuable. This portion is the current core of
Alchemy called, Hg (Mercury).

This is the second iteration of Hg from scratch. I was able to leave
behind some of the constructs that I had originally required for the
first projects goal. Primarily, there is no memory buffer carried around
by default in the current architecture.

I have restructured the files into logical areas to help logically
organize and navigate Network Alchemy. Again, Hg is the only element
that exists in this project currently. Why Hg? Mercury, Messenger of the
Gods.

I am nearing completion of the initial feature set which has been mostly
developed on the VS2013 compiler up to this point. I wanted to clean up
the structure, and ensure the different components compile cleanly for
other platforms.