Commented the inner bit-field classes
Added comments that state design and intent for the inner bit-field
classes.
This is to facilitate the integration of the specialized class to be
added, BitFieldArray.
The BitFieldArray will manage a continuous range of memory, and provide
shift and mask (bit-field) access to each element in the array as if it
were a bit-field.

The addition of this class will allow the implementation of Hg::Array
and Hg::Vector type to be both correct and efficient.