Created bit-field array class specialization
This class presents a bit-field interface to each of the indexed data
items.
The internal memory is still managed as a continuous sequence.
This will simplify serialization and also make it faster.

However, currently there is a trade-off of increased speed of
serialization, and a reduced speed to the access of individual bit-field
values. This issue will be addressed after the features for Milestone 1
are completed.