Integrated Bit-field array class
Support for Arrays of Bit-fields is not supported.

The data is stored in continuous memory similar to how it will be stored
for transport.
Reading and writing the array of data will be fast.
Individually accessing the bit-field interfaces will be slow. This is
logged as an issue and will be revisited for optimization when more
important higher level features are complete.