Interim check in to set progress point
This effort preprocesses the TypeList type defined by the user, and
replaces any appropriate types to create a TypeList suitable for Hg
processing.

Ex. Replace vector<BitField> with BitFieldVector. This allows the
internal processing code to be simplified eliminating many of the
template conditional definitions that were starting to be required.

The status of this branch currently has one failing test, and a few
tests relating to vectors of bit-fields commented out to allow
compiling.