Incremental: Fixed Arrays of Arrays (RayRays)
Fixed read support for arrays of arrays.
Determined that arrays of vectors or arrays of any dynamically sized
sub-type will become overly complex. Therefore I will add tests in the
HgTypeList conversion to convert arrays of dynamically sized sub-types
to vectors of the same type.

Removed the tests for vectors with nested types of mixed size. This is
basically the same as nested with dynamic size.