Corrected reading vectors of dynamic nested fields
- There is one test that fails, which is being kept in place. This is
for arrays of vectors. This will actually be replaced with vectors of
vectors at compile time to simplify the internal logic. However, the
current tests uncovered a defect that prevents identical typelists from
being declared and used separately. This failed test is a reminder to
solve this. There is also a defect logged that details this.

* There is one test for vector reads that is commented out. A mechanism
needs to be created to indicate how many bytes the vector should read
in.

* There are two tests that need to be written to verify that the correct
offsets are being read/written for nested fields.