Reorganized project structure
In order to prepare for support of multiple platforms, the locations of
the project and solution files have been moved.
A build directory has been created, and all of the tests have been moved
to the same sub-directory.
Solution files for earlier versions of Visual Studio will soon be
available.
*nix based platform support with gnu compilers will also be available
soon.