Fixed 64-bit compiler errors.
Discovered that the HG message definition did not match.
This makes me want to get to one of two solutions that will resolve this
issue:
1) Change the MACRO definition table to generate the typelist
internally. This would eliminate the need for a separately defined
typelist. The Bit set automatically generates its own typelist
(integer_sequence template).
2) Add warning tests to the HG_ MACROS that report if a different type
is supplied in the MACRO than was defined in the Typelist. The only
reason the types are listed in the MACRO is to provide a source of
reference for the user that matches the name with the type.