Completed removing the unused offset field.
Completed removing the unused offset field.
This required moving the definition of some of the components to be at
the end of the message definition.
This also requires the name of the structure to be entered in the begin
format MACRO as well as the end format MACRO.
However, this will lead to the ability to internally generate the
Typelist used by a message within the MACRO.
It will also allow the names that are presented to the user to the the
user selected names rather than augmented Hg format names.