Completed basic set of benchmark tests
The current results indicate that Alchemy is 30% slower compared to the
hand-written tests. Therefore, I will return and follow up with some
profiling to improve the performance.

There are no benchmarks for code related to hand-written access of
fields that require shift and mask operations.  This need to be added to
future benchmarks to determine the performance of the Hg::BitList.