Disabling Support for BitField Arrays and Vectors
The previous implementation for packed-bit fields was completely
functional, however, its performance was entirely unacceptable.

The new implementation is much simpler, light-weight, and only about 50%
slower due to necessary internal copies.

However, this also eliminated some of the functionality that was
required these types to work in the BitFieldArrays and BitFieldVectors.
I am temporarily removing these types to allow the basic packed-bit type
to be delivered.

There are other improvements to the library that I believe would provide
more value. Therefore, I will keep this branch available, and attempt to
reintroduce the support for these types in arrays in the future.