Corrected test data for new tests
A more comprehensive test format has been created.
It is a bit pathological to really put the generic template processing
through the ringer.
The test format is a hypothetical structure for describing a 3d geometry
scene. A small collection of triangles, vertices, colors, objects,
instances, transformations and other structures are combined into a
world object.
The world object is what is tested as the data is manipulated.