Interim delivery
Have a better idea of a design direction for a message view object.
This is a handle object that will return iterators for processing.
This removes knowledge of the view objects from the Datum types.

TODO: Complete the initialization functions for msg_view
Create factory functions for msg_view
Integrate into bitmap class
Create external functors for processing pixels with the msg_view
----
At this point add the application logic to collect a message or image to
hide within the carrier image.

Add encryption logic.