Optimization
Converted an internal functor value that was previously copied by-value,
is now copied by-reference.

-This has a significant impact on nested type performance.
-For reasons currently unknown, it dramatically reduces performance on
unaligned memory.

Nested fields will be a far more common scenario than the unaligned
memory field, so I am going with that path.

It may be possible to further optimize to recover the previous
performance for unaligned fields.