Unified the MACRO definitions
Hg and Carbon structures are now both defined through a single set of
"Alchemy" macros.

For all C++ Builds, the Hg structures will be defined.
The Hg classes are not compatible with C.

If the #define ALCHEMY_CARBONATE is defined before the alchemy.h header
is included, then the C-compatible structs will be defined along side
the Hg types.
For non-C++ builds, the C-compatible structs will be defined, but not
the Hg type.