Converted to type aliases and use of auto
The majority of the definitions are now using the template alias syntax
with using.
I have changed a few of the variable declarations to auto where it was
reasonable.
I still need to go through the dense code of pack, unpack, and
byte-order conversion.

As I recognize possibilities, I will add lambdas into these sections.