Optional data type implementation functional
The implementation for the optional data type is fully functional.
However, there are some usability issues that I will be working to
resolve.

Primarily, all data types except automatically flag themselves as valid
if you assign a value to them.

Except for nested data structures when are assigned by sub-fields. The
user can still use the valid() method to flag the optional data
structure as valid. However, this is too error prone to leave this
functionality as it is.