Basic Comparison and Arithmetic Operators
The basic collection of comparison and arithmetic operators are
implemented.
More detailed tests need to better cover the different range of
possibilities, including a range of signs.
The arithmetic operations need to have a complete battery of tests
created that verify all manner of calculations and size expansion.

No optimizations are present (we're a long ways from that)