Added tests for multiplication
There weren't nearly as many possible combinations for multiplication.
As usual, I will need to return with calculations of very large numbers
and stress the memory management.

For now, I would like to get a basic library interface in place to be
able to experiment with the number theory functions.