Power and Squares
Adding additional functionality.
These additional functions are uncovering the additional logic that will
be required in the basic Z integer class.

Will return and add tests once a core collection of additional cases is
collected. I do not want to resolve them now, because the functionality
is not the optimal implementation. I would prefer to see a larger view
of the problems that exist, and factor their solutions where possible.